% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to read/path to that file}
}
\value{
This function returns the selected file as a tbl_df table, or an error if the
file does not exists.
}
\description{
Function that reads a csv file and converts it into a \code{tbl_df} table.
}
\examples{
\dontrun{
fars_read("file.csv")
fars_read("dir/file.csv")
}

}
