% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for a specified year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric or vector with the year(s) used to select the csv files}
}
\value{
This function returns a list with \code{tbl_df} tables containing MONTH and YEAR information for
each selected year or NULL if the \code{year} is not valid.
}
\description{
Function that reads one or more csv files, converts them into \code{tbl_df} tables, extracts year and month
information from these tables and stores the result in a \code{list} object with one entry per year.
}
\details{
The function selects files with names containing one of the year(s) specified by the argument \code{years}.

\code{fars_read_years} is built on the \link{fars_read} and \link{make_filename} functions.
}
\examples{
\dontrun{
fars_read_years(2014)
fars_read_years(2013:2015)[[2]]
}

}
