% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename with a specified year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year to be included in the file name}
}
\value{
This function returns a file name as character string
}
\description{
Function that returns a file name as character string containing a year specified by the argument \code{year}.
Can be used in combination with \link{fars_read}.
}
\examples{
make_filename(2014)
fars_read(make_filename(2015))

}
